feat: redesign landing page with hero, categories, and recent listings
Issue #27: Create a new Landing Page

- Landing page (/) now shows hero section with value proposition and
  CTAs, popular subject categories (max 8 with 'see all' link), and
  recently added book listings
- Gallery+filters moved to /Browse (pre-filtered via ?subject=X)
- Hero uses split layout on desktop (text + preview), single column
  on mobile with preview hidden
- Category tiles use Pico outline buttons with book icon, staggered
  entrance animation honoring prefers-reduced-motion
- Scale on press (0.96) for tactile feedback
- Added tabler-book and tabler-arrow-right icons to sprite
- Edit redirect after delete points to /Browse
