+++ Booker/Pages/Browse.cshtml
@@ -1,0 +1,76 @@
+@page
+@model BrowseModel
+@{
+    ViewData["Title"] = "Przeglądaj ogłoszenia";
+}
+
+<section class="filter">
+    <form novalidate
+          hx-get="/Browse"
+          hx-trigger="submit, input changed delay:500ms, change"
+          hx-target=".grid-gallery"
+          hx-push-url="true">
+        <input type="search" name="search" asp-for="Input.Search" placeholder="Wyszukaj..." data-val="false" hx-on:focus="this.nextElementSibling.style.display = 'block';" hx-on:keydown="this.nextElementSibling.style.display = 'none';" />
+        <small style="display: none;">Szukasz książki do przedmiotu zawodowego lub innej rzadko występującej? Wpisz "inna".</small>
+
+        <details id="filterDetails">
+            <summary id="filterSummary">
+                <svg icon="adjustments-horizontal" aria-hidden="true" style="vertical-align: -0.15em;"></svg>
+                Filtry
+            </summary>
+
+            <div class="grid">
+                <select name="grade" asp-for="Input.Grade" asp-items="Model?.Grades" data-val="false">
+                    <option value="">Wszystkie klasy</option>
+                </select>
+
+                <select name="subject" asp-for="Input.Subject" asp-items="Model?.Subjects" data-val="false">
+                    <option value="">Wszystkie przedmioty</option>
+                </select>
+
+                <select name="level" asp-for="Input.Level" asp-items="Model?.Levels">
+                    <option value="">Wszystkie poziomy</option>
+                </select>
+
+                <input type="number" name="minPrice" asp-for="Input.MinPrice" placeholder="Cena min." data-val="false" />
+                <input type="number" name="maxPrice" asp-for="Input.MaxPrice" placeholder="Cena max." data-val="false" />
+            </div>
+
+            <button type="reset" class="secondary" hx-on:click="this.form.reset(); this.form.dispatchEvent(new Event('submit', { bubbles: true, cancelable: true }));" style="width:100%; margin-top:0.5rem;">Wyczyść filtry</button>
+            <button type="button" id="mobileFilterSubmit" style="width:100%; margin-top:0.5rem;">Zatwierdź filtry</button>
+        </details>
+    </form>
+</section>
+
+<section class="grid-gallery">
+    <vc:item-gallery item-ids="@Model.ItemIds" parameters="@Model.Params" />
+</section>
+
+<section>
+    <div aria-busy="true" class="htmx-indicator" id="ind"></div>
+</section>
+
+<script>
+    function setFilter(name, value, enable) {
+        const elem = document.querySelector(`[name="${name}"]`);
+        if (elem) {
+            elem.value = enable ? value : "";
+            elem.form.requestSubmit();
+        }
+    }
+
+    htmx.onLoad(function(content) {
+        content.querySelectorAll('button.subject, button.grade, button.level').forEach(button => {
+            button.addEventListener('click', function() {
+                const name = this.classList.contains('subject') ? 'subject' : this.classList.contains('grade') ? 'grade' : 'level';
+                const value = name === 'grade' ? this.textContent.trim().slice(6,7) : this.textContent.trim();
+                this.classList.toggle('outline');
+                setFilter(name, value, !this.classList.contains('outline'));
+            });
+        });
+    });
+
+    document.getElementById('mobileFilterSubmit').addEventListener('click', function() {
+        document.getElementById('filterDetails').removeAttribute("open");
+    });
+</script>
