+++ Booker/wwwroot/scss/_landing.scss
@@ -1,0 +1,140 @@
+@use "pico";
+@use "utils";
+@use "sass:map";
+
+/* ===== Hero ===== */
+.landing-hero {
+    display: grid;
+    gap: 2rem;
+    align-items: center;
+    padding-block: clamp(1.5rem, 4vw, 3rem);
+
+    @media (min-width: map.get(map.get(pico.$breakpoints, "lg"), "breakpoint")) {
+        grid-template-columns: 1.1fr 0.9fr;
+    }
+}
+
+.hero-content {
+    h1 {
+        @include utils.h1-size;
+        --pico-font-weight: 800;
+        letter-spacing: -0.02em;
+        margin-bottom: 0.5rem;
+    }
+
+    p {
+        font-size: 1.125rem;
+        line-height: 1.5;
+        margin-bottom: 1.5rem;
+    }
+}
+
+.hero-actions {
+    display: flex;
+    flex-wrap: wrap;
+    gap: 0.75rem;
+
+    a[role="button"] {
+        margin-bottom: 0;
+    }
+}
+
+.hero-preview {
+    [role="article"] {
+        margin-bottom: 0;
+    }
+
+    /* Visually de-emphasize - decorative only */
+    opacity: 0.9;
+
+    @media (max-width: map.get(map.get(pico.$breakpoints, "lg"), "breakpoint") - 1) {
+        display: none;
+    }
+}
+
+/* ===== Sections ===== */
+.landing-section {
+    padding-block: clamp(1.5rem, 3vw, 2.5rem);
+
+    h2 {
+        @include utils.h3-size;
+        margin-bottom: 1.25rem;
+    }
+}
+
+/* ===== Categories ===== */
+.category-grid {
+    list-style: none;
+    margin: 0;
+    padding: 0;
+    display: grid;
+    grid-template-columns: repeat(auto-fill, minmax(150px, 1fr));
+    gap: var(--pico-spacing);
+}
+
+@keyframes categoryEnter {
+    0% {
+        opacity: 0;
+        transform: translateY(0.5rem);
+    }
+
+    100% {
+        opacity: 1;
+        transform: translateY(0);
+    }
+}
+
+.category-tile {
+    animation: categoryEnter 0.3s ease-out backwards;
+    animation-delay: calc(var(--i, 0) * 30ms);
+
+    > a {
+        width: 100%;
+        display: inline-flex;
+        align-items: center;
+        justify-content: center;
+        gap: 0.5rem;
+        min-height: 3rem;
+        text-wrap: balance;
+        text-align: center;
+        font-size: 0.95rem;
+
+        svg {
+            width: 1.125rem;
+            height: auto;
+            flex-shrink: 0;
+        }
+
+        &:active {
+            transform: scale(0.96);
+        }
+    }
+}
+
+.see-all-categories {
+    text-align: center;
+    margin-top: 1.5rem;
+
+    a {
+        display: inline-flex;
+        align-items: center;
+        gap: 0.25rem;
+        color: var(--pico-primary);
+        text-decoration: none;
+        font-weight: 600;
+
+        &:hover {
+            text-decoration: underline;
+        }
+    }
+}
+
+@media (prefers-reduced-motion: reduce) {
+    .category-tile {
+        animation: none;
+    }
+
+    .category-tile > a:active {
+        transform: none;
+    }
+}
